+++ packages/elements/src/input/cad-input.ts
@@ -26,7 +26,7 @@ export type CadInputType =
 export class CadInput extends LitElement {
   static formAssociated = true
 
-  static override shadowRootOptions = {
+  static override shadowRootOptions: ShadowRootInit = {
     ...LitElement.shadowRootOptions,
     delegatesFocus: true,
   }